requirements-server.txt
PyYAML~=5.4
requests~=2.26
tornado~=6.1
mastodon-py~=1.5
matrix-nio~=0.19
